move processes to a new file
