+++ .scripts/start.js
@@ -1,15 +1,9 @@
 #!/usr/bin/env node
 var path = require('path'),
    fs = require('fs'),
-   extend = require('util')._extend,
-   exec = require('child_process').exec,
-   processes = [];
+   processes = require('./processes');
 
-var baseDir = path.resolve(__dirname, '..'),
-   srcDir = path.resolve(baseDir, 'src'),
-   chimpBin = path.resolve(baseDir, '.scripts/node_modules/.bin/chimp'),
-   features = [];
-
+var features = [];
 var argv = require('minimist')(process.argv.slice(2));
 argv._.forEach(function(featureFile) { features.push(path.resolve(featureFile))});
 
@@ -95,7 +89,7 @@ function callChimp(chimpSwitches, callback) {
   if (process.env.WITH_MIRROR) {
     chimpWithMirror(stringify(chimpSwitches));
   } else if (process.env.NO_METEOR) {
-    startChimp('--ddp=' + appOptions.env.ROOT_URL + stringify(chimpSwitches), callback);
+    processes.startChimp('--ddp=' + appOptions.env.ROOT_URL + stringify(chimpSwitches), callback);
   } else {
     // *************************************************
     // THIS IS THE DEFAULT
@@ -117,95 +111,19 @@ function stringify(switches) {
 
 function chimpWithMirror(switches) {
   appOptions.waitForMessage = 'Started MongoDB';
-  startApp(function () {
-    startMirror(function () {
+  processes.startApp(function () {
+    processes.startMirror(function () {
       console.log('=> Test App running at:', mirrorOptions.env.ROOT_URL);
       console.log('=> Log file: tail -f', path.resolve(mirrorOptions.logFile), '\n');
-      startChimp('--ddp=' + mirrorOptions.env.ROOT_URL + switches);
+      processes.startChimp('--ddp=' + mirrorOptions.env.ROOT_URL + switches);
     });
-  });
+  }, appOptions);
 }
 
 function chimpNoMirror(switches) {
   appOptions.waitForMessage = 'App running at';
-  startApp(function () {
+  processes.startApp(function () {
     console.log("inside no mirror ", switches);
-    startChimp('--ddp=' + appOptions.env.ROOT_URL + switches);
-  });
-}
-
-function startApp(callback) {
-  startProcess({
-    name: 'Meteor App',
-    command: 'meteor --settings ' + appOptions.settings + ' --port ' + appOptions.port,
-    waitForMessage: appOptions.waitForMessage,
-    options: {
-      cwd: srcDir,
-      env: extend(appOptions.env, process.env)
-    }
-  }, callback);
-}
-
-function startMirror(callback) {
-  startProcess({
-    // TODO check if settings file exists first
-    name: 'Meteor Mirror',
-    command: 'meteor --settings ' + mirrorOptions.settings + ' --port ' + mirrorOptions.port,
-    silent: true,
-    logFile: mirrorOptions.logFile,
-    waitForMessage: 'App running at',
-    options: {
-      cwd: srcDir,
-      env: extend(mirrorOptions.env, process.env)
-    }
-  }, callback);
-}
-
-function startChimp(command, callback) {
-  console.log("chimpBin ", chimpBin);
-  console.log("command ", command);
-  startProcess({
-    name: 'Chimp',
-    command: chimpBin + ' ' + command,
-    options: {
-      env: Object.assign({}, process.env, {
-        NODE_PATH: process.env.NODE_PATH +
-          path.delimiter + srcDir +
-          path.delimiter + srcDir + '/node_modules',
-      }),
-    },
-  }, callback);
-}
-
-function startProcess(opts, callback) {
-  var proc = exec(
-     opts.command,
-     opts.options
-  );
-  if (opts.waitForMessage) {
-    proc.stdout.on('data', function waitForMessage(data) {
-      if (data.toString().match(opts.waitForMessage)) {
-        if (callback) {
-          callback();
-        }
-      }
-    });
-  }
-  if (!opts.silent) {
-    proc.stdout.pipe(process.stdout);
-    proc.stderr.pipe(process.stderr);
-  }
-  if (opts.logFile) {
-    var logStream = fs.createWriteStream(opts.logFile, {flags: 'a'});
-    proc.stdout.pipe(logStream);
-    proc.stderr.pipe(logStream);
-  }
-  proc.on('close', function (code) {
-    console.log(opts.name, 'exited with code ' + code);
-    for (var i = 0; i < processes.length; i += 1) {
-      processes[i].kill();
-    }
-    process.exit(code);
-  });
-  processes.push(proc);
+    processes.startChimp('--ddp=' + appOptions.env.ROOT_URL + switches);
+  }, appOptions);
 }
